requirements.txt
djangorestframework
djangorestframework-simplejwt
drf-yasg
django-celery-results
celery
redis
twilio
sentence-transformers
faiss-cpu
numpy
h5py
python-dotenv
